+++ src/lib.rs
@@ -29,13 +29,19 @@ macro_rules! base64_serde_type {
     };
     (impl_only, $typename:ident, $config:expr) => {
         impl $typename {
-            pub fn serialize<S>(bytes: &[u8], serializer: S) -> ::std::result::Result<S::Ok, S::Error>
-                where S: $crate::Serializer {
-                serializer.serialize_str(&$crate::encode_config(bytes, $config))
+            pub fn serialize<S, Input>(bytes: Input, serializer: S) -> ::std::result::Result<S::Ok, S::Error>
+                where
+                    S: $crate::Serializer,
+                    Input: AsRef<[u8]>
+            {
+                serializer.serialize_str(&$crate::encode_config(bytes.as_ref(), $config))
             }
 
-            pub fn deserialize<'de, D>(deserializer: D) -> ::std::result::Result<Vec<u8>, D::Error>
-                where D: $crate::Deserializer<'de> {
+            pub fn deserialize<'de, D, Output>(deserializer: D) -> ::std::result::Result<Output, D::Error>
+                where 
+                    D: $crate::Deserializer<'de>,
+                    Output: From<Vec<u8>>
+            {
                 struct Base64Visitor;
 
                 impl<'de> $crate::de::Visitor<'de> for Base64Visitor {
@@ -51,7 +57,7 @@ macro_rules! base64_serde_type {
                     }
                 }
 
-                deserializer.deserialize_str(Base64Visitor)
+                deserializer.deserialize_str(Base64Visitor).map(|vec| Output::from(vec))
             }
         }
     };