+++ RELEASE-NOTES.md
@@ -1,3 +1,8 @@
+# 0.6.0
+
+- Now serializer works with any input type that implements `AsRef<[u8]>`
+- Now deserializer works with any output type that implements `From<Vec<u8>>`
+
 # 0.5.0
 
 - Use `base64` 0.12.0